+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### Added
 - `Properties::load_command`.
 - `Line::TargetYawOverride`.
+- `HLTAS::to_static()`.
 
 ### Changed
 - Changed all `&'a str` types into `Cow<'a, str>` thus making it possible to construct a new `HLTAS` programmatically with owned strings.